requirements.txt
altair==4.2.2
anyio==4.0.0
argon2-cffi-bindings==21.2.0
argon2-cffi==23.1.0
argparse==1.4.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==22.2.0
babel==2.13.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.5
bs4==0.0.2
cachecontrol==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.2.0
configparser==5.3.0
cryptography==39.0.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
firebase-admin==6.1.0
firebase-auth==3.0.28
fqdn==1.5.1
gcloud==0.17.0
gitdb==4.0.10
gitpython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-auth==2.16.0
google-cloud-access-approval==1.12.0
google-cloud-access-context-manager==0.1.16
google-cloud-api-keys==0.5.1
google-cloud-appengine-logging==1.3.0
google-cloud-asset==3.20.1
google-cloud-audit-log==0.2.5
google-cloud-billing==1.12.0
google-cloud-compute==1.10.1
google-cloud-core==2.3.2
google-cloud-dataproc==5.4.0
google-cloud-dns==0.34.2
google-cloud-error-reporting==1.8.2
google-cloud-essential-contacts==1.5.0
google-cloud-firestore==2.9.1
google-cloud-functions==1.11.0
google-cloud-iam==2.11.2
google-cloud-kms==2.15.0
google-cloud-logging==3.5.0
google-cloud-monitoring-metrics-scopes==1.4.4
google-cloud-monitoring==2.16.0
google-cloud-org-policy==1.8.3
google-cloud-os-config==1.15.3
google-cloud-recaptcha-enterprise==1.11.1
google-cloud-resource-manager==1.8.1
google-cloud-run==0.10.0
google-cloud-secret-manager==2.16.0
google-cloud-securesourcemanager==0.1.1
google-cloud-service-usage==1.8.0
google-cloud-storage==2.7.0
google-cloud==0.34.0
google-crc32c==1.5.0
google-oauth==1.0.1
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio-helpers==0.0.9
grpcio-status==1.51.1
grpcio-tools==1.51.3
grpcio==1.51.3
httplib2==0.21.0
hydraml==0.0.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.26.0
ipython==8.12.3
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema-specifications==2023.7.1
jsonschema==4.19.2
jupyter-client==8.6.0
jupyter-console==6.6.3
jupyter-core==5.5.0
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter-server-terminals==0.4.4
jupyter-server==2.10.0
jupyter==1.0.0
jupyterlab-pygments==0.2.2
jupyterlab-server==2.25.0
jupyterlab-widgets==3.0.9
jupyterlab==4.0.8
jws==0.1.3
lxml==5.1.0
markdown-it-py==2.2.0
markupsafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
msgpack==1.0.4
nbclient==0.8.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook-shim==0.2.3
notebook==7.0.6
numpy==1.24.2
oauth2client==3.0.0
oauthlib==3.2.2
openpyxl==3.1.1
overrides==7.4.0
packaging==23.0
pamqp==2.3.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pika==1.3.2
pillow==9.4.0
pip==23.3.1
pkgutil-resolve-name==1.3.10
platformdirs==3.11.0
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf3==0.2.1
protobuf==4.22.1
psutil==5.9.6
pulp==2.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.10.1
pydeck==0.8.0
pygments==2.14.0
pyjwt==2.6.0
pympler==1.0.1
pyopenssl==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-jwt==2.0.1
pytz-deprecation-shim==0.1.0.post0
pytz==2022.7.1
pywin32==306
pywinpty==2.0.12
pyyaml==6.0.1
pyzmq==25.1.1
qtconsole==5.5.0
qtpy==2.4.1
referencing==0.30.2
regex==2022.10.31
render==1.0.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.1
rpds-py==0.12.0
rsa==4.9
semver==2.13.0
send2trash==1.8.2
setuptools==65.5.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
streamlit==1.19.0
tabulate==0.9.0
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.14
validators==0.20.0
watchdog==2.3.0
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
werkzeug==2.2.3
wheel==0.38.4
widgetsnbextension==4.0.9
zipp==3.14.0
